Major fixes
The variable "play" has been removed, now there is another method that uses css to disable clicking on the deck.

Instead of manually creating an array with duplicate values, now im using concat to repeat them.

The "InitEventSystem" has been removed, its methods are now replaced with "setTimeout" to wait for the end of the animations and the click listener is initialized on the "InitDeck".

The popup was not showing after the first time.